Implemented item backend item deletingfunction in itemModule files: itemController, itemModel, itemRoute
